+++ packages/driver/src/dom/visibility.ts
@@ -206,11 +206,15 @@ const elHasOverflowHidden = function ($el) {
   return cssOverflow.includes('hidden')
 }
 
-const elHasPositionRelative = ($el) => {
+const elHasPositionRelative = ($el: JQuery<HTMLElement>) => {
   return $el.css('position') === 'relative'
 }
 
-const elHasPositionAbsolute = ($el) => {
+const elHasPositionStatic = ($el: JQuery<HTMLElement>) => {
+  return $el.css('position') == null || $el.css('position') === 'static'
+}
+
+const elHasPositionAbsolute = ($el: JQuery<HTMLElement>) => {
   return $el.css('position') === 'absolute'
 }
 
@@ -220,13 +224,12 @@ const elHasClippableOverflow = function ($el) {
             OVERFLOW_PROPS.includes($el.css('overflow-x'))
 }
 
-const canClipContent = function ($el, $ancestor) {
+const canClipContent = function ($el: JQuery<HTMLElement>, $ancestor: JQuery<HTMLElement>) {
   // can't clip without overflow properties
   if (!elHasClippableOverflow($ancestor)) {
     return false
   }
 
-  // fix for 29605 - display: contents
   if (elHasDisplayContents($ancestor)) {
     return false
   }
@@ -248,8 +251,18 @@ const canClipContent = function ($el, $ancestor) {
 
   // even if ancestors' overflow is clippable, if the element's offset parent
   // is a child of the ancestor, the ancestor will not clip the element
-  // unless the ancestor has position absolute
+  // unless the ancestor has a position that is not absolute
   if (elHasPositionAbsolute($offsetParent) && isChild($ancestor, $offsetParent)) {
+    return false
+  }
+
+  // even if ancestors' overflow is clippable,
+  // if the element is position static or relative,
+  // and the element's offset parent is positioned absolute, a descendent of the ancestor, and has no clippable overflow,
+  // then the ancestor will not clip the element
+  if ((elHasPositionStatic($el) || elHasPositionRelative($el))
+    && elHasPositionAbsolute($offsetParent) && isDescendent($ancestor, $offsetParent) && !elHasClippableOverflow($offsetParent)
+  ) {
     return false
   }
 
@@ -266,8 +279,7 @@ export const isW3CFocusable = (el) => {
   return isFocusable(wrap(el)) && isW3CRendered(el)
 }
 
-// @ts-ignore
-const elAtCenterPoint = function ($el) {
+const elAtCenterPoint = function ($el: JQuery<HTMLElement>) {
   const doc = $document.getDocumentFromElement($el.get(0))
   const elProps = $coordinates.getElementPositioning($el)
 
@@ -278,6 +290,8 @@ const elAtCenterPoint = function ($el) {
   if (el) {
     return $jquery.wrap(el)
   }
+
+  return undefined
 }
 
 const elDescendentsHavePositionFixedOrAbsolute = function ($parent, $child) {
@@ -298,7 +312,7 @@ const elHasVisibleChild = function ($el) {
   })
 }
 
-const elIsNotElementFromPoint = function ($el) {
+const elIsNotElementFromPoint = function ($el: JQuery<HTMLElement>) {
   // if we have a fixed position element that means
   // it is fixed 'relative' to the viewport which means
   // it MUST be available with elementFromPoint because
@@ -333,7 +347,6 @@ const elIsOutOfBoundsOfAncestorsOverflow = function ($el: JQuery<any>, $ancestor
     return false
   }
 
-  // fix for 29605 - display: contents
   if (elHasDisplayContents($el)) {
     return false
   }
@@ -400,8 +413,7 @@ const elIsHiddenByAncestors = function ($el, checkOpacity, $origEl = $el) {
   }
 
   if (elHasOverflowHidden($parent) && !elHasDisplayContents($parent) && elHasNoEffectiveWidthOrHeight($parent)) {
-    // if any of the elements between the parent and origEl
-    // have fixed or position absolute
+    // if any of the elements between the parent and origEl have fixed or position absolute
     return !elDescendentsHavePositionFixedOrAbsolute($parent, $origEl)
   }
 
@@ -564,7 +576,6 @@ export const getReasonIsHidden = function ($el, options = { checkOpacity: true }
     return `This element \`${node}\` is not visible because its parent \`${parentNode}\` has CSS property: \`overflow: hidden\` and an effective width and height of: \`${width} x ${height}\` pixels.`
   }
 
-  // nested else --___________--
   if (elOrAncestorIsFixedOrSticky($el)) {
     if (elIsNotElementFromPoint($el)) {
       // show the long element here